refactoring code, ckeditor etc.

+++ service/resources/js/Helpers/js/axios.js
@@ -26,15 +26,27 @@ export const axiosPost = async (request) => {
 
 export const axiosDelete = async (request) => {
 	try {
-		const response = await window.axios.delete(
-			request.url,
-			{ data: { id: request.id } },
-			{
+		let response
+		if (Object.prototype.hasOwnProperty.call(request, "id")) {
+			response = await window.axios.delete(
+				request.url,
+				{ data: { id: request.id } },
+				{
+					headers: {
+						"Content-Type": "application/json"
+					}
+				}
+			)
+		}
+		if (Object.prototype.hasOwnProperty.call(request, "json")) {
+			response = await window.axios.delete(request.url, {
+				data: request.json,
 				headers: {
 					"Content-Type": "application/json"
 				}
-			}
-		)
+			})
+		}
+		if (!response) throw new Error("❌ Cannot delete. No id, json, type or date")
 		return response.data
 	} catch (error) {
 		axiosError(200, error)
